+++ tests/OpenIdConnectTest.php
@@ -2,6 +2,7 @@
 
 namespace yiiunit\extensions\authclient;
 
+use yii\authclient\OAuthToken;
 use yii\authclient\OpenIdConnect;
 use yii\caching\ArrayCache;
 
@@ -32,6 +33,9 @@ class OpenIdConnectTest extends TestCase
         $this->assertTrue(isset($configParams['token_endpoint']));
 
         $this->assertEquals($configParams['token_endpoint'], $authClient->getConfigParam('token_endpoint'));
+
+        // Test default value for non existing
+        $this->assertEquals('default', $authClient->getConfigParam('non-existing', 'default'));
     }
 
     /**
@@ -83,5 +87,70 @@ class OpenIdConnectTest extends TestCase
         $this->assertNotEmpty($authClient->authUrl);
         $this->assertContains($clientId, $builtAuthUrl, 'No client id present!');
         $this->assertContains(rawurlencode($returnUrl), $builtAuthUrl, 'No return URL present!');
+    }
+
+    public function testNonce()
+    {
+        $authClient = new OpenIdConnect([
+            'issuerUrl' => 'https://accounts.google.com',
+            'cache' => null,
+        ]);
+        $clientId = 'test_client_id';
+        $authClient->clientId = $clientId;
+        $returnUrl = 'http://test.return.url';
+        $nonce = 'test_nonce';
+        $code = 'test_code';
+        $authClient->setReturnUrl($returnUrl);
+
+        $builtAuthUrl = $authClient->buildAuthUrl([
+            'nonce' => $nonce,
+        ]);
+
+
+        $query = parse_url($builtAuthUrl)['query'];
+        parse_str($query, $query_vars);
+        $this->assertEquals($query_vars['nonce'], $nonce);
+    }
+
+    public function testUserInfoFromToken()
+    {
+        $accessToken = new OAuthToken([
+            'params' => [
+                'id_token' => 'eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiJ0ZXN0LWNsaWVudC10eXBlLWF1dGgtY29kZS1vcGVuLWlkLWNvbm5lY3QiLCJpc3MiOiJodHRwczovL2xvY2FsaG9zdCIsImlhdCI6MTYzNTY5NDUzNi40MjkyMzcsImV4cCI6NDc5MTM2ODEzNiwic3ViIjoiMTIzIiwiYXV0aF90aW1lIjoxNjM1NjkwOTM1LCJub25jZSI6InNWbEVmS2xNMTdhYlJfY2Q5cXhvcU5fZHN3a0VRWXUxIn0.LgV-jFoopYnEhgygtt4bDL4HV1Rnw_cuKgopQ_I8f2YxDUlXKO2M0ANjA1iWIsBTCAKnI5JF7wYWBlK7eFJkU16U8yNVYHyUNaMGzXG1Q3khLmPfa9tmU2Kj2loA2hGGkZTjHCAuDgYSSFucLlFnqcR4-vhhwUyZdvFvwRRi0FF1r10m2oNmzfVLAcQxo2C5C_inSmuGnzfWqvrsDjdnT8N2XE2e3hVRxlIEv4GkupUejjdyWlSBjsUjnfXMlmi6VBn7HfElcVjJqp3L7GHVV1zfSA82e3oo7_wvQbb090M4nwFOmasvGnvZddELQdxL9KW0s_AIdkUM5lFxFFnl8Q',
+            ]
+        ]);
+
+        $authClient = new OpenIdConnect([
+            'cache' => null,
+            'configParams' => [],
+            'accessToken' => $accessToken,
+            'validateJws' => false,
+        ]);
+
+        $userAttributes = $authClient->getUserAttributes();
+
+        $this->assertEquals(['sub' => '123'], $userAttributes);
+    }
+
+    public function testUserInfoFromUserInfoTokenResponse()
+    {
+        /** @var OpenIdConnect $oidcClient */
+        $oidcClient = $this->getMockBuilder(OpenIdConnect::className())
+            ->setMethods(['api'])
+            ->getMock();
+
+        $oidcClient->expects($this->once())
+            ->method('api')
+            ->willReturn(
+                'eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiJ0ZXN0LWNsaWVudC10eXBlLWF1dGgtY29kZS1vcGVuLWlkLWNvbm5lY3QiLCJpc3MiOiJodHRwczovL2xvY2FsaG9zdCIsImlhdCI6MTYzNTY5NDUzNi40MjkyMzcsImV4cCI6NDc5MTM2ODEzNiwic3ViIjoiMTIzIiwiYXV0aF90aW1lIjoxNjM1NjkwOTM1LCJub25jZSI6InNWbEVmS2xNMTdhYlJfY2Q5cXhvcU5fZHN3a0VRWXUxIn0.LgV-jFoopYnEhgygtt4bDL4HV1Rnw_cuKgopQ_I8f2YxDUlXKO2M0ANjA1iWIsBTCAKnI5JF7wYWBlK7eFJkU16U8yNVYHyUNaMGzXG1Q3khLmPfa9tmU2Kj2loA2hGGkZTjHCAuDgYSSFucLlFnqcR4-vhhwUyZdvFvwRRi0FF1r10m2oNmzfVLAcQxo2C5C_inSmuGnzfWqvrsDjdnT8N2XE2e3hVRxlIEv4GkupUejjdyWlSBjsUjnfXMlmi6VBn7HfElcVjJqp3L7GHVV1zfSA82e3oo7_wvQbb090M4nwFOmasvGnvZddELQdxL9KW0s_AIdkUM5lFxFFnl8Q'
+            );
+
+        $oidcClient->cache = null;
+        $oidcClient->configParams = ['userinfo_endpoint' => 'http://localhost'];
+        $oidcClient->validateJws = false;
+
+        $userAttributes = $oidcClient->getUserAttributes();
+
+        $this->assertEquals(['sub' => '123'], $userAttributes);
     }
 }